requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
motor>=3.0.0
pymongo>=3.12.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
pydantic>=1.8.2
python-dateutil>=2.8.2
email-validator>=1.1.3
aioredis>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
pydantic-settings>=2.0.0
pydantic>=2.0.0
redis>=4.5.0